+++ NEWS.md
@@ -1,9 +1,10 @@
-# crochet 2.0.0.9000
+# crochet 2.0.1
 
 * Fix incorrect index type determination when additional arguments were passed
   to `extract()` and `replace()`.
 * Fix missing `...` pass-through for single indices in `extract()`.
 * Throw an error if index type is not supported.
+
 
 # crochet 2.0.0